services/image-processing-service/requirements.txt
flask
gevent
pillow
requests
jsonpickle
numpy
opencv-contrib-python